MJ Changes
Modification to the code to allow proper positioning of the nozzle. Still working on resetting the encoder count. the encoder is going backwards?? idk why.

MAX

+++ include/mjdriver.h
@@ -44,7 +44,7 @@ public:
     void power_off();
 
     void report_status();
-    void set_absolute_start(int steps);
+    void set_absolute_start(double steps);
     void set_printing_frequency(int frequency);
     void set_head_voltage(HeadIndex idx, double voltage);
     void mode_select(Mode mode);
